# Fast path for varint read.